refactor(fleet): consolidate Bridge package and simplify formation wiring
- Unify Agent Panel and Carrier UI overlays under bridge/:
  bridge/acp-shell (ACP overlay shell), bridge/panel (panel state/lifecycle),
  bridge/render (panel rendering engine), bridge/streaming (stream store),
  bridge/carrier-ui (Status Overlay, TaskForce Config Overlay, status renderer,
  overlay frame/model-flow). bridge/index.ts reduces to a bootBridge facade.
- Split overlay-specific types (CarrierCliType, CarrierStatusEntry, CliModelInfo,
  OverlayState, CarrierOverlayCallbacks, etc.) from shipyard/carrier/types.ts
  into bridge/carrier-ui/types.ts so shipyard retains only carrier-framework
  primitives (CarrierConfig, CarrierMetadata, CARRIER_FRAMEWORK_KEY).
- Drop callback-based formation notification (onSortieStateChange,
  onSquadronStateChange, onTaskForceConfigChange) and inline pi.registerTool
  ownership into fleet/index.ts; buildSortieToolConfig / buildTaskForceToolConfig
  / buildSquadronToolConfig are invoked directly at boot and at state transitions.
- Update consumers: fleet/index.ts, bridge/panel/widget-sync.ts, overlay tests,
  stream-store/footer-cols/block-renderer/message-renderers tests, and AGENTS.md
  (bridge directory tree, dependency principles, module structure table).
