+++ extensions/fleet/bridge/acp-shell/command.ts
