+++ extensions/fleet/bridge/carrier-ui/status-overlay.ts
