+++ extensions/fleet/bridge/panel/config.ts
@@ -6,8 +6,8 @@
  */
 
 import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
-import { MIN_BODY_H, MAX_BODY_H, formatPanelMultiColHint } from "../constants";
-import type { ServiceSnapshot } from "../../core/agentclientprotocol/types.js";
+import { MIN_BODY_H, MAX_BODY_H, formatPanelMultiColHint } from "../../constants";
+import type { ServiceSnapshot } from "../../../core/agentclientprotocol/types.js";
 import { getState } from "./state.js";
 import { syncWidget } from "./widget-sync.js";
