+++ extensions/fleet/bridge/panel/types.ts
