+++ extensions/fleet/bridge/render/compact.ts
