+++ extensions/fleet/bridge/streaming/types.ts
@@ -7,7 +7,7 @@
  * ⚠️ 이 파일은 런타임 코드가 아닌 순수 타입/인터페이스만 포함합니다.
  */
 
-import type { AgentStatus } from "../../core/agentclientprotocol/types.js";
+import type { AgentStatus } from "../../../core/agentclientprotocol/types.js";
 
 // ─── 스트리밍 블록 타입 ──────────────────────────────────
