+++ extensions/fleet/shipyard/carrier/sortie.ts
@@ -25,8 +25,8 @@ import { truncateToWidth, visibleWidth } from "@mariozechner/pi-tui";
 import { getLogAPI } from "../../../core/log/bridge.js";
 import { runAgentRequest } from "../../operation-runner.js";
 import { composeTier2Request } from "./prompts.js";
-import { getVisibleRun, getRunById } from "../../streaming/stream-store.js";
-import { renderBlockLines, blockLineToAnsi } from "../../render/block-renderer.js";
+import { getVisibleRun, getRunById } from "../../bridge/streaming/stream-store.js";
+import { renderBlockLines, blockLineToAnsi } from "../../bridge/render/block-renderer.js";
 import {
   getRegisteredOrder,
   getSortieEnabledIds,
@@ -141,8 +141,8 @@ const SORTIE_RESULT_CACHE_KEY = "__pi_carrier_sortie_result_cache__";
 /**
  * carriers_sortie 도구 정의(ToolDefinition)를 조립해 반환합니다.
  *
- * pi.registerTool 호출 오너쉽은 admiral/prompts.ts의 `registerSortieTool`이
- * 소유합니다. 이 팩토리는 등록 시 필요한 schema/guidelines/execute/render 등
+ * pi.registerTool 호출 오너쉽은 fleet/index.ts가 부팅 시 1회 등록합니다.
+ * 이 팩토리는 등록 시 필요한 schema/guidelines/execute/render 등
  * 도구 기능 자체만을 제공합니다. 등록 불필요 시 null을 반환합니다.
  */
 export function buildSortieToolConfig() {
