+++ extensions/fleet/tests/stream-store.test.ts
@@ -17,7 +17,7 @@ import {
   getAllVisibleRuns,
   getRunById,
   resetRuns,
-} from "../streaming/stream-store.js";
+} from "../bridge/streaming/stream-store.js";
 
 // 각 테스트 전에 globalThis 상태 초기화
 beforeEach(() => {
